Four folders added
Git offers a suite of collaboration tools to make working with others on a project easier. In order to collaborate we need :
•	A complete replica of the project on our own computers
•	A way to keep track of and review each other’s work
•	Access to a definitive project version
We can accomplish all of this by using remotes. A remote is a shared Git repository that allows multiple collaborators to work on the same Git project from different locations. Collaborators work on the project independently, and merge changes together when they are ready to do so.